@@ -16,7 +16,7 @@ Gem::Specification.new do |s|
 
   s.files = Dir["{app,config,db,lib}/**/*", "MIT-LICENSE", "Rakefile", "README.md"]
 
-  s.add_dependency "rails", "~> 5.1"
+  s.add_dependency "rails", "~> 6.0.0"
   s.add_dependency 'mongoid', "~> 7.3.0"
   s.add_runtime_dependency 'mongoid-paperclip'
   s.add_runtime_dependency 'mongoid_paranoia'
