@@ -8,12 +8,15 @@ Ishapi::Engine.routes.draw do
   get 'cities/view/:cityname', :to => 'cities#show'
   get 'cities/features',       :to => 'cities#features'
 
+  ## @TODO: remove all co-tailors stuff
+=begin
   post 'co_tailors/orders',                 :to => 'orders#create'
   post 'co_tailors/order_items',            :to => 'order_items#create'
   post 'co_tailors/measurements',           :to => 'measurements#update'
   get  'co_tailors/products/by-kind/:kind', :to => 'products#show'
   namespace :co_tailors do
   end
+=end
 
   get 'events/view/:eventname',      :to => 'events#show'
 
@@ -27,8 +30,7 @@ Ishapi::Engine.routes.draw do
   get 'maps', to: 'maps#index'
   get 'maps/view/:slug', to: 'maps#show'
   get 'markers/view/:slug', to: 'maps#show_marker'
-  get  "/my/account", to: "users#account"
-  post "/my/account", to: "users#account"
+  match  "/my/account", to: "users#account", via: [ :get, :post ]
   namespace :my do
     get 'galleries', to: 'galleries#index'
     get 'newsitems', to: 'newsitems#index'
@@ -49,24 +51,28 @@ Ishapi::Engine.routes.draw do
   get 'reports', :to => 'reports#index'
   get 'reports/view/:slug', :to => 'reports#show'
 
-  get 'sites/view/:domain',                            :to => 'sites#show',      :constraints => { :domain => /[^\/]+/ }
-  post 'sites/view/:domain',                           :to => 'sites#show',      :constraints => { :domain => /[^\/]+/ }
-  get 'sites/view/:domain/newsitems/:newsitems_page',  :to => 'newsitems#index', :constraints => { :domain => /[^\/]+/ }
-  get 'sites/view/:domain/reports',                    :to => 'reports#index',   :constraints => { :domain => /[^\/]+/ }
-  get 'sites/view/:domain/reports/page/:reports_page', :to => 'reports#index',   :constraints => { :domain => /[^\/]+/ }
-  get 'sites/view/:domain/tags',                       :to => 'tags#index',      :constraints => { :domain => /[^\/]+/ }
+  get  'sites/view/:domain',                            :to => 'sites#show',      :constraints => { :domain => /[^\/]+/ }
+  post 'sites/view/:domain',                            :to => 'sites#show',      :constraints => { :domain => /[^\/]+/ }
+  get  'sites/view/:domain/newsitems/:newsitems_page',  :to => 'newsitems#index', :constraints => { :domain => /[^\/]+/ }
+  get  'sites/view/:domain/reports',                    :to => 'reports#index',   :constraints => { :domain => /[^\/]+/ }
+  get  'sites/view/:domain/reports/page/:reports_page', :to => 'reports#index',   :constraints => { :domain => /[^\/]+/ }
+  get  'sites/view/:domain/tags',                       :to => 'tags#index',      :constraints => { :domain => /[^\/]+/ }
 
   post 'stars/buy', to: 'gameui#buy_stars'
+
+  ## 2022-02-12 moved to iron_warbler gem _vp_
+  # resources "stock_watches"
 
   get 'tags/view/:slug', :to => 'tags#show'
   get 'test', to: 'application#test'
 
-  post  'users/fb_sign_in',     :to => 'users#fb_sign_in'
-  post  'users/profile',        :to => 'users#show'
-  post  'users/profile/update', :to => 'users#update'
-  get   'users/profile',        :to => 'users#show' # @TODO: only for testing! accessToken must be hidden
+  post  'users/fb_sign_in',      to: 'users#fb_sign_in'
+  get   'users/me',              to: 'users#account'
+  post  'users/profile',         to: 'users#show'
+  post  'users/profile/update',  to: 'users#update'
+  get   'users/profile',         to: 'users#show' # @TODO: only for testing! accessToken must be hidden
   match 'users/long_term_token', to: 'application#long_term_token', via: [ :get, :post ]
-  post  'users/login', to: 'users#login'
+  post  'users/login',           to: 'users#login'
 
   get 'venues', :to => 'venues#index'
   get 'venues/view/:venuename', :to => 'venues#show'
